Created Initialization infrastructure https://github.com/achobanov/web-application/issues/3
