+++ Gateways/Web.Contracts/Initialization.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace Blog.Gateways.Web.Contracts
+{
+    public interface IInitializtion
+    {
+        Task Initialize(IServiceProvider serviceProvider);
+    }
+}
