+++ Gateways/Web/Infrastructure/Extensions/WebHostExtensions.cs
@@ -1,0 +1,24 @@
+﻿using System.Linq;
+using System.Threading.Tasks;
+using Blog.Gateways.Web.Contracts;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Blog.Gateways.Web.Infrastructure.Extensions
+{
+    public static class WebHostExtensions
+    {
+        public static async Task Initialize(this IWebHost webHost)
+        {
+            using var scope = webHost.Services.CreateScope();
+            var services = scope.ServiceProvider;
+
+            var initializations = webHost.Services.GetServices<IInitializtion>();
+
+            foreach (var initalization in initializations)
+            {
+                await initalization.Initialize(services);
+            }
+        }
+    }
+}
