+++ Gateways/Web/Program.cs
@@ -1,3 +1,5 @@
+using System.Threading.Tasks;
+using Blog.Gateways.Web.Infrastructure.Extensions;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 
@@ -5,11 +7,14 @@ namespace Blog.Gateways.Web
 {
     public class Program
     {
-        public static void Main(string[] args) 
-            => CreateWebHostBuilder(args)
-                .Build()
-                .Initialize()
-                .Run();
+        public static async Task Main(string[] args)
+        {
+            var webHost = CreateWebHostBuilder(args).Build();
+
+            await webHost.Initialize();
+
+            await webHost.RunAsync();
+        }
 
         public static IWebHostBuilder CreateWebHostBuilder(string[] args)
             => WebHost
