+++ Gateways/Web/WebServices.cs
@@ -1,4 +1,7 @@
-﻿using Blog.Common.ConventionalServices;
+﻿using System;
+using System.Linq;
+using System.Reflection;
+using Blog.Common.ConventionalServices;
 using Blog.Gateways.Web.Providers;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -11,5 +14,25 @@ namespace Blog.Gateways.Web
             => services
                 .AddHttpContextAccessor()
                 .AddContractProviders(typeof(AuthenticationProvider).Assembly);
+
+        public static IServiceCollection AddContractProviders(
+            this IServiceCollection services,
+            params Assembly[] assemblies)
+        {
+            var contractProviderType = typeof(IContractProvider);
+
+            var contractProviders = assemblies
+                .SelectMany(assembly => assembly.GetExportedTypes())
+                .Where(type => type.IsClass && !type.IsAbstract && contractProviderType.IsAssignableFrom(type))
+                .Select(Activator.CreateInstance)
+                .Cast<IContractProvider>();
+
+            foreach (var provider in contractProviders)
+            {
+                provider.ProvideImplementations(services);
+            }
+
+            return services;
+        }
     }
 }
